validação de usuário e token, página principal transferida, logs no institutions e campuses
