+++ pages/api/v1/bonds/index.ts
@@ -12,32 +12,47 @@ type RequestBody = {
 export type BondsResponse = { data: IBondDTOProps[] };
 export default async function Bonds(
   request: NextApiRequest,
-  response: NextApiResponse< BondsResponse| { error: string }>
+  response: NextApiResponse<BondsResponse | { error: string }>
 ) {
   logger.log("Bonds", "Request received", {});
-  const { username, institution, token } = JSON.parse(JSON.stringify(request.body)) as RequestBody;
+  const { username, institution, token } = JSON.parse(
+    JSON.stringify(request.body)
+  ) as RequestBody;
   if (!token) return response.status(400).send({ error: "Token is required" });
-  if (!institution) return response.status(400).send({ data:undefined, error: "Institution is required" });
+  if (!institution)
+    return response
+      .status(400)
+      .send({ data: undefined, error: "Institution is required" });
 
-  const compatibleInstitution = compatibleInstitutions.find(i => i.acronym === institution);
-  if (!compatibleInstitution) return response.status(400).send({ data:undefined, error: "Institution is not compatible" });
+  const compatibleInstitution = compatibleInstitutions.find(
+    (i) => i.acronym === institution
+  );
+  if (!compatibleInstitution)
+    return response
+      .status(400)
+      .send({ data: undefined, error: "Institution is not compatible" });
 
   const sigaaURL = compatibleInstitution.url;
 
   logger.log("Bonds", "Token received", token);
   const storedSession = await prisma.session.findUnique({
-    where: { token },
+    where: {
+      token,
+      Student: {
+        username
+      },
+    },
     select: { value: true },
   });
   if (!storedSession)
-    return response.status(400).send({ error: "Invalid token" });
+    return response.status(400).send({ error: "Invalid token or username" });
   logger.log("Bonds", "JSESSIONID received", storedSession.value);
   const authService = new AuthService();
   const accountService = await authService.rehydrate({
     JSESSIONID: storedSession.value,
     username,
     url: sigaaURL,
-    institution: compatibleInstitution.acronym
+    institution: compatibleInstitution.acronym,
   });
   logger.log("Bonds", "Account service rehydrated", {});
   const activeBonds = await accountService.getActiveBonds();
